// src/modules/solicitud/enums/solicitud.enum.ts
export enum EstadoSolicitud {
  RECEPCIONADA = 'RECEPCIONADA',
  APROBADA = 'APROBADA',
  DEVUELTA = 'DEVUELTA',
  INFORME_EMITIDO = 'INFORME_EMITIDO',
  INSTALADA = 'INSTALADA',
  EN_CONTROL = 'EN_CONTROL',
  NO_FACTIBLE = 'NO_FACTIBLE',
  NO_RECOMENDABLE = 'NO_RECOMENDABLE',
  CERRADA = 'CERRADA',
  ANULADA = 'ANULADA',
}

export enum OrigenCreacion {
  FORMULARIO_WEB = 'FORMULARIO_WEB',
  INTERCONEXION_PJUD = 'INTERCONEXION_PJUD',
}

export enum TipoCausa {
  RUC_RIT = 'RUC_RIT',
  ROL = 'ROL',
}

export enum MotivoOrigen {
  ORIGINAL = 'ORIGINAL',
  CAMBIO_DOMICILIO = 'CAMBIO_DOMICILIO',
}